% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fivetexts}
\alias{fivetexts}
\title{five texts}
\format{
{A data frame with two columns and five rows \describe{
\item{\code{id}}{character} \item{\code{text}}{character} } }
}
\usage{
fivetexts
}
\description{
Five texts combining titles and abstracts about the Bahamas from
the Lens database
}
\details{
Search of the \href{https://www.lens.org/}{Lens scholarly database}
\url{https://www.lens.org/lens/search/scholar/list?q=bahamas}
}
\examples{
fivetexts
}
\keyword{datasets}
