% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnfinder.R
\name{gnfinder}
\alias{gnfinder}
\title{Find names with gnfinder}
\usage{
gnfinder(
  string,
  id = NULL,
  nobayes = NULL,
  check_names = NULL,
  sources = NULL,
  source_ids = NULL
)
}
\arguments{
\item{string}{A string or strings to be queried for a species name (character)}

\item{id}{A document identifier (e.g. df$id). If NULL added as 1:n}

\item{nobayes}{Whether to turn off the bayes algorithm (default = FALSE)}

\item{check_names}{check identified names against taxonomic databases?
(logical). Defaults to TRUE for the Catalogue of Life}

\item{sources}{flag for sources}

\item{source_ids}{list of taxonomic databases. See list at
\url{https://index.globalnames.org/datasource}
(integer) * 1 = Catalogue of Life * 2 = Wikispecies * 3 = ITIS * 9 = World
Register of Marine Species * 11 = GBIF Backbone Taxonomy * 12 =
Encyclopedia of Life * 179 = Open Tree of Life Reference Taxonomy}
}
\value{
data.frame
}
\description{
Extract names from a text by passing a string or strings and
returning a data frame. Optionally check names and retrieve taxon ids and
classification from different data sources.
}
\examples{
single <- gnfinder("Lepidium meyenii is a hot plant. Escherichia coli is not")
hotplants <- gnfinder(c("Lepidium meyenii is a hot plant",
"Capsicum annuum is a hot plant for a different reason"))
five <- gnfinder(fivetexts$text, fivetexts$id, check_names = FALSE)
}
