% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onehundredtexts}
\alias{onehundredtexts}
\title{one hundred texts}
\format{
{A data frame with two columns and 100 rows \describe{
\item{\code{id}}{character} \item{\code{text}}{character} } }
}
\usage{
onehundredtexts
}
\description{
One hundred texts containing combined titles and abstracts about
the Bahamas from the Lens database
}
\details{
Search of the \href{https://www.lens.org/}{Lens scholarly database}
\url{https://www.lens.org/lens/search/scholar/list?q=bahamas}
}
\examples{
onehundredtexts
}
\keyword{datasets}
