% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zootaxa}
\alias{zootaxa}
\title{zootaxa titles}
\format{
{A data frame with two columns and 27974 rows \describe{
\item{\code{paperid}}{numeric Microsoft Academic Graph identifies}
\item{\code{doi}}{character} \item{\code{originaltitle}}{character} } }
}
\usage{
zootaxa
}
\description{
Titles from the Zootaxa journal sourced from Microsoft Academic
Graph (May 2020) including the DOI and MAG paperid.
}
\details{
Data from \href{https://aka.ms/msracad}{Microsoft Academic Graph} under an
Open Data Commons Attribution License (ODC-By)
\url{https://opendatacommons.org/licenses/by/1-0/}
}
\examples{
zootaxa
}
\keyword{datasets}
